+++ src/main/java/org/squad9/vehiclerentalservice/service/ManufacturerServiceImpl.java
@@ -1,11 +1,17 @@
 package org.squad9.vehiclerentalservice.service;
 
 import lombok.AllArgsConstructor;
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
+import org.squad9.vehiclerentalservice.dto.request.ManufacturerRequestDTO;
+import org.squad9.vehiclerentalservice.dto.response.AccessoryResponseDTO;
+import org.squad9.vehiclerentalservice.dto.response.ManufacturerResponseDTO;
+import org.squad9.vehiclerentalservice.model.AccessoryModel;
 import org.squad9.vehiclerentalservice.model.ManufacturerModel;
 import org.squad9.vehiclerentalservice.repository.ManufacturerRepository;
 import org.squad9.vehiclerentalservice.service.interfaces.ManufacturerService;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -13,60 +19,51 @@ import java.util.UUID;
 @Service
 @AllArgsConstructor
 public class ManufacturerServiceImpl implements ManufacturerService {
-    private ManufacturerRepository manufacturerRepository;
+    private final ManufacturerRepository manufacturerRepository;
+    private final ModelMapper modelMapper;
 
     @Override
-    public ManufacturerModel save(ManufacturerModel fabricante) {
-        try {
-            return manufacturerRepository.save(fabricante);
-        }catch (Exception e){
-            throw new RuntimeException(e.getMessage());
-        }
+    public ManufacturerResponseDTO save(ManufacturerRequestDTO request) {
+        ManufacturerModel manufacturerToSave = modelMapper.map(request, ManufacturerModel.class);
+        ManufacturerModel savedManufacturer = manufacturerRepository.save(manufacturerToSave);
+
+        return modelMapper.map(savedManufacturer, ManufacturerResponseDTO.class);
     }
 
     @Override
-    public List<ManufacturerModel> findAll() {
-        try {
-            return manufacturerRepository.findAll();
-        }catch (Exception e) {
-            throw new RuntimeException(e.getMessage());
-        }
+    public List<ManufacturerResponseDTO> findAll() {
+        List<ManufacturerModel> manufacturers = manufacturerRepository.findAll();
+        List<ManufacturerResponseDTO> response = new ArrayList<>();
+
+        manufacturers.forEach(manufacturer -> response.add(modelMapper.map(manufacturer, ManufacturerResponseDTO.class)));
+        return response;
     }
 
     @Override
-    public ManufacturerModel findById(UUID id){
-        try{
-            Optional<ManufacturerModel> fabricanteOptional = manufacturerRepository.findById(id);
-            if (fabricanteOptional.isPresent()){
-                return fabricanteOptional.get();
-            }
-        } catch (Exception e) {
-            throw new RuntimeException(e.getMessage());
-        }
-        return null;
+    public ManufacturerResponseDTO findById(UUID id){
+        ManufacturerModel manufacturer = manufacturerRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Fabricante não encontrado com o ID: " + id));
+
+        return modelMapper.map(manufacturer, ManufacturerResponseDTO.class);
     }
 
     @Override
     public void remove(UUID id){
-        try {
-            manufacturerRepository.deleteById(id);
-        } catch (Exception e) {
-            throw new RuntimeException("Erro ao remover fabricante: " + e.getMessage());
+        if (!manufacturerRepository.existsById(id)) {
+            throw new IllegalArgumentException("Fabricante não encontrado com o ID: " + id);
         }
+        manufacturerRepository.deleteById(id);
     }
 
     @Override
-    public ManufacturerModel update(UUID id, ManufacturerModel manufacturer) {
-        try {
-            ManufacturerModel existingManufacturer = manufacturerRepository.findById(id)
-                    .orElseThrow(() -> new IllegalArgumentException("Fabricante não encontrado com o ID: " + id));
+    public ManufacturerResponseDTO update(UUID id, ManufacturerRequestDTO request) {
+        manufacturerRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Acessório não encontrado com o ID: " + id));
 
-            existingManufacturer.setName(manufacturer.getName());
+        ManufacturerModel manufacturerToUpdate = modelMapper.map(request, ManufacturerModel.class);
+        manufacturerToUpdate.setId(id);
+        ManufacturerModel updatedManufacturer = manufacturerRepository.save(manufacturerToUpdate);
 
-            return manufacturerRepository.save(existingManufacturer);
-        } catch (Exception e) {
-            throw new RuntimeException("Erro ao atualizar fabricante: " + e.getMessage());
-        }
+        return modelMapper.map(updatedManufacturer, ManufacturerResponseDTO.class);
     }
 
 }
